requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
msal>=1.24.0
PyJWT>=2.8.0
cryptography>=41.0.0
tqdm>=4.66.0
